Assistant checkpoint: Added comprehensive logging and fixed syntax error
Assistant generated file changes:
- server/routes.ts: Enhanced image search endpoint with logging, Fix imports structure

---

User prompt:

This is what my app is trying to achieve....  we are trying to take the description of the Yorkie provided by the user, and then use the Discord Bot which we have setup and added to the channel where all the Yorkie pictures are, to find a few upscaled pictures of Yorkies which we recommend to the user. Then we take the image that the user selected and send it to OpenAi via the API to get a description of the Yorkshire terrier and a story based on it, along with the other story details which we have previously collected from the user.  Once we receive the reply back from OpenAi we present the user with a storybook page with the picture they selected and the story that was created.  I feel like this is not being handled well and I have very little insight into why things are failing