Assistant checkpoint: Add image search function and fix OpenAI model name
Assistant generated file changes:
- server/lib/openai.ts: Fix OpenAI model name
- server/lib/discord.ts: Add findSimilarYorkieImage function

---

User prompt:

we are trying to take the description of the Yorkie provided by the user, and then use the Discord Bot which we have setup and added to the channel where all the Yorkie pictures are, to find a few upscaled pictures of Yorkies which we recommend to the user. Then we take the image that the user selected and send it to OpenAi via the API to get a description of the Yorkshire terrier and a story based on it, along with the other story details which we have previously collected from the user.  Once we receive the reply back from OpenAi we present the user with a storybook page with the picture they selected and the story that was created.  do you see where the app is failing?